+++ .env.example
@@ -56,3 +56,7 @@ VITE_PUSHER_HOST="${PUSHER_HOST}"
 VITE_PUSHER_PORT="${PUSHER_PORT}"
 VITE_PUSHER_SCHEME="${PUSHER_SCHEME}"
 VITE_PUSHER_APP_CLUSTER="${PUSHER_APP_CLUSTER}"
+
+SEARCH_URL=https://googleapis.com/customsearch/v1
+SEARCH_API_KEY=取得したAPI_Keyを設定
+SEARCH_CX=取得した検索IDを設定
